feat: extract pixi_compute_sources from the dispatcher
