+++ crates/pixi_command_dispatcher/src/util/mod.rs
@@ -1,6 +1,5 @@
 pub mod executor;
 pub(crate) mod limits;
-pub(crate) mod path;
 pub(crate) mod ptr_arc;
 
 pub use executor::Executor;
